third_party/blink/web_tests/wpt_internal/battery-status/multiple-promises.html

<!DOCTYPE html>
<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/services/device/public/mojom/battery_monitor.mojom.js"></script>
<script src="resources/mock-battery-monitor.js"></script>
<script>
promise_test(async _ => {
  const promise1 = navigator.getBattery();
  const promise2 = navigator.getBattery();
  assert_equals(promise1, promise2);

  setAndFireMockBatteryInfo(false, 10, 20, 0.5);
  testIfBatteryStatusIsUpToDate(await promise1);
  testIfBatteryStatusIsUpToDate(await promise2);

}, 'multiple parallel invocations of navigator.getBattery()');
</script>
</body>
</html>
